+++ examples/common/Hello.ts
@@ -1,3 +1,0 @@
-export interface Hello {
-  hello: string;
-}